requirements.txt
Pillow
tqdm==4.67.1
accelerate==1.0.1
transformers==4.46.3
scipy==1.10.1
dreamsim==0.2.1